refactor: 역할분리 리팩토링
