+++ app/src/main/java/nextstep/jwp/infrastructure/RequestMapper.java
@@ -1,5 +1,6 @@
 package nextstep.jwp.infrastructure;
 
+import nextstep.jwp.infrastructure.processor.LoginRequestProcessor;
 import nextstep.jwp.infrastructure.processor.RequestProcessor;
 
 import java.util.HashMap;
@@ -9,4 +10,11 @@ public class RequestMapper {
 
     private final static Map<String, RequestProcessor> MAPPER = new HashMap<>();
 
+    static {
+        MAPPER.put("/login", new LoginRequestProcessor());
+    }
+
+    public static RequestProcessor mappingProcessor(String path) {
+        return MAPPER.get(path);
+    }
 }
