+++ app/src/main/java/nextstep/jwp/infrastructure/processor/RequestProcessor.java
@@ -1,4 +1,10 @@
 package nextstep.jwp.infrastructure.processor;
 
+import nextstep.jwp.model.CustomHttpRequest;
+
+import java.io.OutputStream;
+
 public interface RequestProcessor {
+
+    public String processResponse(CustomHttpRequest request, OutputStream outputStream);
 }
