+++ app/src/main/java/nextstep/jwp/model/ContentType.java
@@ -1,0 +1,13 @@
+package nextstep.jwp.model;
+
+public enum ContentType {
+    HTML("text/html"),
+    CSS("text/css"),
+    JS("text/js");
+
+    String contentType;
+
+    ContentType(String contentType) {
+        this.contentType = contentType;
+    }
+}
